// type.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
type Review = {
    rating: number;
    comment: string;
    date: string;
    reviewerName: string;
    reviewerEmail: string;
  };
  
  type Meta = {
    createdAt: string;
    updatedAt: string;
    barcode: string;
    qrCode: string;
  };
  
  type Dimensions = {
    width: number;
    height: number;
    depth: number;
  };
  
  export type ProductType = {
    id: number;
    title: string;
    description: string;
    category: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    tags: string[];
    brand: string;
    sku: string;
    weight: number;
    dimensions: Dimensions;
    warrantyInformation: string;
    shippingInformation: string;
    availabilityStatus: string;
    reviews: Review[];
    returnPolicy: string;
    minimumOrderQuantity: number;
    meta: Meta;
    images: string[];
    thumbnail: string;
    quantity?: number;
  };


  export interface StateType {
    shofy:{
      cart: ProductType[];
      favorite: ProductType[];
      userInfo: any;
    }
  }
  